backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
dnspython==2.4.2
whois==0.9.15
requests==2.31.0
aiohttp==3.9.1
reportlab==4.0.7
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2
pytz==2023.3
jinja2==3.1.2

